requirements.txt
et-xmlfile==1.1.0
lxml==5.3.0
openpyxl==3.1.5
pillow==10.4.0
python-pptx==1.0.2
typing_extensions==4.12.2
XlsxWriter==3.2.0
